requirements.txt
numpy>=1.20.0
torch>=1.9.0
geoopt>=0.5.0
matplotlib>=3.4.0
seaborn>=0.11.0
scikit-learn>=0.24.0
pandas>=1.3.0
tqdm>=4.60.0
Flask>=2.0
biopython>=1.79
peptides>=0.3.0
